docker_simulator/simulator/requirements.txt
pandas
kafka-python
pyarrow
